requirements.txt
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.23
pypdfium2==4.30.0
setuptools==80.9.0
zplgrf==1.6.0
